Converted .js files to .tsx and added typing for image assets
