--- src/assets/index.js
@@ -1,2 +1,0 @@
-export {default as CICSoftLogo} from './img/webp/cicsoft-logo.webp';
-export {default as FallbackCICSoftLogo} from './img/png/cicsoft-logo.png';
