+++ src/assets/index.tsx
@@ -1,0 +1,2 @@
+export { default as CICSoftLogo } from "./img/webp/cicsoft-logo.webp";
+export { default as FallbackCICSoftLogo } from "./img/png/cicsoft-logo.png";
