+++ src/common.d.ts
@@ -1,0 +1,19 @@
+declare module "*.png" {
+  const path: string;
+  export default path;
+}
+
+declare module "*.jpg" {
+  const path: string;
+  export default path;
+}
+
+declare module "*.webp" {
+  const path: string;
+  export default path;
+}
+
+declare module "*.mp4" {
+  const path: string;
+  export default path;
+}
