+++ src/utils/Image.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+type ImageProps = {
+  src: string;
+  alt: string;
+  srcSet: string;
+  [key: string]: any;
+};
+
+export default function Image({ src, alt, srcSet, ...props }: ImageProps) {
+  return (
+    <picture>
+      <source srcSet={srcSet} type="image/webp" />
+      <img src={src} alt={alt} {...props} />
+    </picture>
+  );
+}
